System updated to version 0.303 Components updated to latest corresponding versions

+++ components/modules/Plupload/trigger/installed.php
@@ -32,14 +32,8 @@ Trigger::instance()->register(
 		while ($f = $cdb->fs($files)) {
 			$storage->unlink($f);
 		}
-		$storage->rmdir('Plupload');
-		clean_pcache();
-	}
-);
-if (!function_exists(__NAMESPACE__.'\\clean_pcache')) {
-	function clean_pcache () {
-		if (file_exists(PCACHE.'/module.Plupload.js')) {
-			unlink(PCACHE.'/module.Plupload.js');
+		if ($storage->is_dir('Plupload')) {
+			$storage->rmdir('Plupload');
 		}
 	}
-}
+);
